📝 🐛 update Swagger
