+++ 2NotyBackend/controllers/subscribers.js
@@ -1,3 +1,4 @@
+const { pool } = require("./../dbCongif");
 const getSubscribers = async (req, res = response) => {
     pool.connect().then((client) => {
       return client
